(#14584) Add option to squelch standard error
Some Facter resolutions will attempt to call commands that may not exist
on certain platforms or the commands may issue warning to STDERR. These
warnings and errors are not relevant to Facter, nor do they affect the
operation of Facter. Prior to this commit, resolution writers were
required to add '2>/dev/null' to any command executed to prevent
warnings or errors from showing in Facter's output. This commit adds the
ability to add the `squelch_standard_error` to the fact resolution to
accomplish the same result.  Additionally, the output to STDERR isn't
thrown away, but can be acted upon programmatically within the
resolution, but merely doesn't show to the end user.